refactor: moved auth components to page folder
